crawler/requirements.txt
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
elasticsearch==7.10.1
future==0.18.2
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
parsel==1.6.0
pep8==1.7.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pymongo==3.11.2
pyOpenSSL==20.0.1
python-dotenv==0.15.0
queuelib==1.5.0
Scrapy==2.4.1
scrapyd==1.2.1
service-identity==18.1.0
six==1.15.0
toml==0.10.2
Twisted==20.3.0
urllib3==1.26.3
w3lib==1.22.0
zope.interface==5.2.0
